idp-cdk-app/src/lambda/requirements.txt
boto3==1.24.61
amazon-textract-caller==0.0.24
amazon-textract-response-parser==0.1.33
amazon-textract-prettyprinter==0.0.16
amazon-textract-overlayer==0.0.10
xlsxwriter==3.0.3
filetype
Pillow
pdfplumber